+++ R/offspring.R
@@ -5,7 +5,7 @@
 ##' @examples
 ##' library(ape)
 ##' tree <- rtree(4)
-##' x <- as_data_frame(tree)
+##' x <- as_tibble(tree)
 ##' child(x, 4)
 child.tbl_tree <- function(.data, .node, ...) {
     valid.tbl_tree(.data)
@@ -28,7 +28,7 @@ child.tbl_tree <- function(.data, .node, ...) {
 ##' @examples
 ##' library(ape)
 ##' tree <- rtree(4)
-##' x <- as_data_frame(tree)
+##' x <- as_tibble(tree)
 ##' offspring(x, 4)
 offspring.tbl_tree <- function(.data, .node, ...) {
     x <- child(.data, .node)